requirements.txt
fastapi
uvicorn[standard]
streamlit
tensorflow
pillow
requests
python-multipart
numpy